Split server code and main code
